devdep/workspace/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
slowapi==0.1.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pytest==8.2.0
httpx==0.26.0
pytest-asyncio==1.3.0
